+++ tests/test_reactions.py
@@ -1,5 +1,5 @@
-from autode.reactions.reaction_types import classify
-from autode.exceptions import ReactionFormationFalied
+from autode.reactions.reaction_types import classify, Addition
+from autode.exceptions import ReactionFormationFailed
 import pytest
 
 
@@ -23,11 +23,20 @@ def test_classify():
     assert rearrangement.name == 'rearrangement'
 
     # Needs to have at least some reactants and products
-    with pytest.raises(ReactionFormationFalied):
+    with pytest.raises(ReactionFormationFailed):
         _ = classify([], [])
+
+    with pytest.raises(ReactionFormationFailed):
         _ = classify([0], [])
+
+    with pytest.raises(ReactionFormationFailed):
         _ = classify([], [0])
 
     # 3 -> 3 reactions are not currently supported
     with pytest.raises(NotImplementedError):
         _ = classify([0, 1, 2], [3, 4, 5])
+
+
+def test_equality():
+
+    assert Addition == Addition